# ideas and structures/openapi.frontend.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DAMN 前端所需的后端 API",
    "version": "0.2.0",
    "description": "自适应任务规划器前端（frontend/）当前依赖与将要依赖的全部 HTTP 接口。\n\n约定：\n- API 前缀 `/api/v1`，健康检查在根路径 `/health`，**不在** `/api/v1` 下。\n- 除 `/health`、`/auth/*` 外都需要 `Authorization: Bearer <JWT>`。\n- 所有错误统一返回 `{ code, message, detail }`。\n- 每个操作带 `x-implementation-status`：`existing` = 后端已有；`to-add` = 需要后端新增。\n- 日期用 `YYYY-MM-DD`；时间用 `HH:MM:SS`；时间戳用 ISO 8601。",
    "contact": { "name": "前端 | docs/design/backend-api-gaps.md" }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "开发环境（局域网）"
    }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/goals": {
      "get": {
        "tags": ["goals"],
        "summary": "目标列表 / 待拆解草稿",
        "description": "目标页「待拆解清单」跨设备保留。`status=draft` 为待拆解项。不传 status 返回全部。",
        "x-implementation-status": "to-add",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["active", "draft", "completed", "archived", "cancelled"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "目标列表",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/GoalDetailRead" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["goals"],
        "summary": "新建目标 / 待拆解项",
        "description": "`status=draft` 表示一条等待拆解的任务；`active` 表示已确认进行中的目标。",
        "x-implementation-status": "to-add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/GoalCreate" } }
          }
        },
        "responses": {
          "201": {
            "description": "新目标",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/GoalDetailRead" } }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/Validation" }
        }
      }
    },
    "/api/v1/goals/{goal_id}": {
      "patch": {
        "tags": ["goals"],
        "summary": "更新目标 / 待拆解项",
        "description": "编辑待拆解项的名称、优先级、截止日期、补充。",
        "x-implementation-status": "to-add",
        "parameters": [{ "$ref": "#/components/parameters/GoalId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/GoalUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的目标",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/GoalDetailRead" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["goals"],
        "summary": "删除目标 / 待拆解项",
        "x-implementation-status": "to-add",
        "parameters": [{ "$ref": "#/components/parameters/GoalId" }],
        "responses": {
          "204": { "description": "已删除" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
...
      "GoalId": {
        "name": "goal_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    },
    "responses": {
...
      "NotFound": {
        "description": "资源不存在（404）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      },
      "Conflict": {
        "description": "冲突（409）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      },
      "Validation": {
        "description": "参数校验失败（422）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "description": "统一错误体",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "description": "机器可读错误码，如 `not_found` / `conflict` / `replan_not_eligible` / `validation_error`"
          },
          "message": { "type": "string", "description": "给人看的说明" },
          "detail": { "nullable": true, "description": "可选的补充信息（对象或数组）" }
        }
      },
...
      "GoalDetailRead": {
        "type": "object",
        "description": "目标 / 待拆解项",
        "required": ["id", "title", "status"],
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true, "description": "待拆解项里对应「补充」" },
          "goal_type": { "type": "string", "default": "other" },
          "status": {
            "type": "string",
            "enum": ["active", "draft", "completed", "archived", "cancelled"],
            "description": "draft = 等待拆解"
          },
          "deadline": { "type": "string", "format": "date", "nullable": true },
          "priority": { "type": "integer", "minimum": 1, "maximum": 4 },
          "estimated_minutes": { "type": "integer", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "GoalCreate": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "goal_type": { "type": "string", "enum": ["long_term", "short_term", "project", "habit", "other"], "default": "other" },
          "deadline": { "type": "string", "format": "date-time", "nullable": true },
          "priority": { "type": "integer", "minimum": 1, "maximum": 4, "default": 2 },
          "estimated_minutes": { "type": "integer", "nullable": true },
          "subject": { "type": "string", "nullable": true },
          "task_type": { "type": "string", "nullable": true },
          "status": {
            "type": "string",
            "enum": ["active", "draft"],
            "default": "active",
            "description": "建待拆解项时传 draft"
          }
        }
      },
      "GoalUpdate": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "status": {
            "type": "string",
            "enum": ["active", "draft", "completed", "archived", "cancelled"]
          },
          "deadline": { "type": "string", "format": "date", "nullable": true },
          "priority": { "type": "integer", "minimum": 1, "maximum": 4 }
        }
      },
...
    }
  }
}